Add placeholder image that is displayed when direct link is used for image that does not exist
I started working on this for #13, to provide a placeholder image thumbnail in comments section in instances where user commented on deleted image. Then I split this off due to scope, as well as the fact that this would be a good feature to have in not just comments section, but also for direct link accesses too. It seems like imgur takes this approach as well, and that's what I based my implementation off of.

Other Changes:
- Created integration test utils, which contains just `addImagesForUser` right now. Modified User Images integ test to use this util function as well and ensured those tests still pass.
- Added 500 status code error response if DB encountered error while finding image via direct link, an error response would be returned in place of the placeholder image in this case